Add files via upload
In this case study, I created a sales dashboard in Excel that can help to analyse the store’s sales performance. Also, it should help improve work management across the different selling platforms and best selling products. It is the great way to visualize the workload and team’s performance in terms of how many orders are being processed and how quickly the store delivers its products.